feat(refactoring): timer and timer validators improved and fixed, cleanup, husky pre-commit improved
