+++ package.json
@@ -12,6 +12,7 @@
     "lint": "ng lint",
     "lint:fix": "ng lint --fix",
     "format": "prettier --write \"**/*.{ts,html,scss,json,md}\"",
+    "postinstall": "husky",
     "prepare": "husky"
   },
   "private": true,
